apps/biz-api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
alembic==1.16.4
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
python-jose==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pandas==2.3.1
openpyxl==3.1.5
google-api-python-client==2.176.0
google-auth==2.40.3
pytest==8.4.1
